% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.R
\name{E}
\alias{E}
\title{Expected Values}
\usage{
E(c, lambda, mu, p0, nmax, type = "")
}
\arguments{
\item{c}{number of servers}

\item{lambda}{arrival rate}

\item{mu}{service rate}

\item{p0}{probability of having 0 subjects/objects in the system (has to be calculated before)}

\item{nmax}{max number of subjects/objects allowed in the system}

\item{type}{character, type in E_Nq, E_N, E_Wq, E_W for the desired expected value}
}
\value{
Expected Values either of E_Nq, E_N, E_Wq, E_W
}
\description{
Returns the expected value depending on the entry for type
}
