% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pn.R
\name{Pn}
\alias{Pn}
\title{Pn Function}
\usage{
Pn(c, lambda, mu, nmax, n)
}
\arguments{
\item{c}{number of servers}

\item{lambda}{arrival rate}

\item{mu}{service rate}

\item{nmax}{max number of subjects/objects allowed in the system}

\item{n}{number of subjects/objects in the system}
}
\value{
Probability
}
\description{
Probability Pn of having n customers in the System
}
